Add support for random item names
Adds the function item_name_generator which allows random naming of items.

The function currently generates a random name from two lists of strings.

item_name_generator accepts an item as an argument.

In future updates, the function will look at the item's attributes and generate a name that is based on the attributes of the item.