Add support for random warrior and item names
Adds support for randomly generated warrior and item names.

Functions item_name_generator and warrior_name_generator both accept one argument and return a string.

The string is currently generated from lists, but in future versions of these functions the name will be generated based on the attributes of the object passed into the function.

This update also changes the file structure to group the Item and Warrior code together.

Adds an arena_code folder. This is where the code that runs when the fight occurs will be stored.