Add run_simulation function
The run_simulation function will be used to make the warriors fight.

It accepts the group of warriors, and will run modular and customizable functions imported from their own files to make the warriors fight.

The intention is to allow for greater customization and tweaking of the logic, so the simulation can develop in complexity without becoming impossible to debug.

ALSO adds more flair to the item and warrior names.