Adds warrior_setup and item_setup and cleans up comments
The lists of warriors and items should be generated by function calls rather than logic in the main app file.

This update is the first step in tackling a refactor.

In future updates, the app logic will be built out inside the app_logic folder, and multiple function calls that mutate data will be used to perform the simulation.