Add very basic support for fight mechanics
Adds the barebones minimum needed to have each Warrior in the warriors array deal damage to a menacing Ogre.

The run_simulation function loops through the warriors list and has each one deal damage according to their ap to the Ogre.

If the Ogre's hp falls to 0 or below, the program prints out that the Ogre is dead and the name of the warrior who killed him.

In future updates the warriors will fight each other, and the attacks will happen asynchronously.