Combat Update - Add combat to simulation
It's here!

The app will now simulate combat between two random warriors.

Note: this version only supports even numbered lists of warriors. Odd numbered lists will cause problems.

The run_simulation function now loops through the list of warriors passed in and pops two random warriors off into variables called "comb1" and "comb2" (combatant 1 and combatant 2)

It then calls the warrior's fight method until one of their hp stats falls to or below 0.

It puts the winning warrior in a list of winners, and prints out the list of winners at the end of combat.

BONUS: also added support in the run_simulation function for the winner to pick up the items the loser was carrying.

In future versions, this should be refactored to be a method on the Warrior class.

That's it! There's combat now! Holy crap it works!

This is my greatest achievement - a functional MVP in one week in Python, which I have never used before this Monday.

I can't wait to build this sim out more as an exercise in learning Python. Perhaps someday I'll even deploy it...