V0.2 - Tier Support Update
The MVP is complete!

This is the big one.

Added so much stuff. I got a huge burst of energy and decided to just crank it out.

New in this update:
- Updated combat logic
- Game not terminated in either one final winner or, if all combatants die, no winner.
- Warrior stats are now affected when they pick up an item
- Changed default Warrior hp stat range
- Warriors can now pick up a random item from a list, and pick up a specific item. Both methods modify the Warrior's base ap stat

This update contains all the features I intended to create when I started this learning process.

Now, I intend to take some Python classes, and continue to develop this app - perhaps as a full-stack React app deployed as a portfolio piece...