V 0.5 - App Logic Overhaul
This update contains huge updates to the app logic, including:

- new Custom_Warrior class
- new "commence_battle" function
- major changes to the loops in "run_simulation"
- new imports throughout
- tweaks to the Warrior and Custom_Warrior class methods
- new "run_simulation_v2" function
- unused "custom_items" file slated for future development

NOTE: this update has bugs. Lots of em.

The entire update needs to be combed through. There's a lot of stuff that can be improved here, and some app-breaking bugs.

BUT

IT WORKS :)

You can generate as many Warriors and Items as you want. Warriors and Items are programatically generated and then they all fight to the death, until only one winner remains!

This is the foundation of me:
a) learning how to code in Python
b) creating a mini evolution simulator

More on that later, but for now, the core feature of the app, a grand battle royale, works.